Explicitly set CORS headers
By default FastAPI only replies with CORS headers when an origin is
sent. This means that the response should be sent with Vary: Origin,
but it isn't. If this header were added it would reduce cachability.

When we always send the headers this means that the response can be
cached regardless of origin, and CORS are always present.