Switch to layered storage
This substantially alters the in-DB storage format, breaking each
tile into layers. In the future this will allow for tiles to be
refreshed per-layer, reducing excess work. A tile database will
need to be re-created after this change.

It also substantially improves SQL building and escaping, along
with better test coverage of Storage